Split eigenvalue solvers into separate implementations
